Extracted common geomtry styling to abstract viewmodel
