+++ app/src/main/java/com/teamagam/gimelgimel/app/map/view/DynamicLayerActionFragment.java
@@ -35,8 +35,7 @@ public class DynamicLayerActionFragment
     View view = super.onCreateView(inflater, container, savedInstanceState);
     mApp.getApplicationComponent().inject(this);
 
-    mViewModel = mViewModelFactory.create(getContext(), mGGMapView, super::pickColor,
-        super::pickBorderStyle);
+    mViewModel = mViewModelFactory.create(mGGMapView, super::pickColor, super::pickBorderStyle);
     mViewModel.init();
 
     setBottombarListeners();
