+++ app/src/main/java/com/teamagam/gimelgimel/app/map/viewModel/DynamicLayerEditViewModel.java
@@ -2,15 +2,16 @@ package com.teamagam.gimelgimel.app.map.viewModel;
 
 import android.content.Context;
 import android.databinding.Observable;
-import android.graphics.Color;
 import android.view.View;
+import com.android.databinding.library.baseAdapters.BR;
 import com.google.auto.factory.AutoFactory;
 import com.google.auto.factory.Provided;
 import com.teamagam.gimelgimel.R;
 import com.teamagam.gimelgimel.app.map.view.GGMapView;
 import com.teamagam.gimelgimel.app.map.viewModel.gestures.OnMapGestureListener;
+import com.teamagam.gimelgimel.domain.dynamicLayers.DisplayDynamicLayersInteractorFactory;
+import com.teamagam.gimelgimel.domain.dynamicLayers.remote.SendRemoteAddDynamicEntityRequestInteractorFactory;
 import com.teamagam.gimelgimel.domain.icons.DisplayIconsInteractorFactory;
-import com.teamagam.gimelgimel.domain.icons.entities.Icon;
 import com.teamagam.gimelgimel.domain.layers.DisplayVectorLayersInteractorFactory;
 import com.teamagam.gimelgimel.domain.map.DisplayMapEntitiesInteractorFactory;
 import com.teamagam.gimelgimel.domain.map.entities.geometries.PointGeometry;
@@ -27,95 +28,58 @@ import java.util.ArrayList;
 import java.util.List;
 
 @AutoFactory
-public class DynamicLayerEditViewModel extends BaseMapViewModel {
+public class DynamicLayerEditViewModel extends BaseGeometryStyleViewModel {
 
-  private static final String DEFAULT_BORDER_STYLE = "solid";
-
+  private final SendRemoteAddDynamicEntityRequestInteractorFactory
+      mAddDynamicEntityRequestInteractorFactory;
   private final List<PointGeometry> mPoints;
   private final List<GeoEntity> mNewEntities;
-  private final DisplayIconsInteractorFactory mDisplayIconsInteractorFactory;
   private GGMapView mGGMapView;
   private MapDrawer mMapDrawer;
   private MapEntityFactory mEntityFactory;
-  private OnDrawGestureListener mGestureLisetner;
+  private OnDrawGestureListener mGestureListener;
   private Consumer<PointGeometry> mOnMapClick;
-  private Consumer<Integer> mPickColor;
-  private Consumer<String> mPickBorderStyle;
   private GeoEntity mCurrentEntity;
-  private String mBorderColor;
-  private String mBorderStyle;
-  private String mFillColor;
-  private List<Icon> mIcons;
-  private int mTypeIdx;
   private boolean mIsOnEditMode;
-  private boolean mIsBorderColorPicking;
   private int mBorderStyleVisibility;
   private int mBorderColorVisibility;
   private int mFillColorVisibility;
   private int mIconPickerVisibility;
 
-  protected DynamicLayerEditViewModel(
+  protected DynamicLayerEditViewModel(@Provided Context context,
       @Provided DisplayMapEntitiesInteractorFactory displayMapEntitiesInteractorFactory,
       @Provided DisplayVectorLayersInteractorFactory displayVectorLayersInteractorFactory,
+      @Provided DisplayDynamicLayersInteractorFactory displayDynamicLayersInteractorFactory,
       @Provided
           DisplayIntermediateRastersInteractorFactory displayIntermediateRastersInteractorFactory,
       @Provided DisplayIconsInteractorFactory displayIconsInteractorFactory,
-      Context context,
+      @Provided
+          SendRemoteAddDynamicEntityRequestInteractorFactory addDynamicEntityRequestInteractorFactory,
       GGMapView ggMapView,
       Consumer<Integer> pickColor,
       Consumer<String> pickBorderStyle) {
     super(displayMapEntitiesInteractorFactory, displayVectorLayersInteractorFactory,
-        displayIntermediateRastersInteractorFactory, ggMapView);
+        displayDynamicLayersInteractorFactory, displayIntermediateRastersInteractorFactory,
+        displayIconsInteractorFactory, context, ggMapView, pickColor, pickBorderStyle);
 
+    mAddDynamicEntityRequestInteractorFactory = addDynamicEntityRequestInteractorFactory;
     mGGMapView = ggMapView;
-    mDisplayIconsInteractorFactory = displayIconsInteractorFactory;
-    mPickColor = pickColor;
-    mPickBorderStyle = pickBorderStyle;
 
     mIsOnEditMode = false;
-    mIsBorderColorPicking = false;
-    mIcons = new ArrayList<>();
     mPoints = new ArrayList<>();
     mNewEntities = new ArrayList<>();
     mMapDrawer = new MapDrawer(mGGMapView);
     mEntityFactory = new MapEntityFactory();
-    mGestureLisetner = new OnDrawGestureListener();
-    mBorderColor = colorToString(context.getResources().getColor(R.color.default_border_color));
-    mBorderStyle = DEFAULT_BORDER_STYLE;
-    mFillColor = colorToString(context.getResources().getColor(R.color.default_fill_color));
+    mGestureListener = new OnDrawGestureListener();
 
     mGGMapView.setOnMapGestureListener(null);
-
     addOnPropertyChangedCallback(new OnPropertyChanged());
 
-    mDisplayIconsInteractorFactory.create(icon -> {
-      mIcons.add(icon);
-    }).execute();
+    super.loadIcons();
   }
 
   public boolean isOnEditMode() {
     return mIsOnEditMode;
-  }
-
-  public String[] getTypes() {
-    return generateIconNames();
-  }
-
-  public int getTypeIdx() {
-    return mTypeIdx;
-  }
-
-  public void setTypeIdx(int typeId) {
-    mTypeIdx = typeId;
-    notifyPropertyChanged(0);
-  }
-
-  public int getBorderColor() {
-    return Color.parseColor(mBorderColor);
-  }
-
-  public int getFillColor() {
-    return Color.parseColor(mFillColor);
   }
 
   public int getBorderStyleVisibility() {
@@ -144,7 +108,9 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
     mIsOnEditMode = false;
 
     if (mCurrentEntity != null) {
+      mAddDynamicEntityRequestInteractorFactory.create(mCurrentEntity).execute();
       mNewEntities.add(mCurrentEntity);
+      mMapDrawer.erase(mCurrentEntity);
       clearCurrentEntity();
     }
   }
@@ -153,49 +119,6 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
     mIsOnEditMode = false;
     mMapDrawer.erase(mCurrentEntity);
     clearCurrentEntity();
-  }
-
-  public void onBorderStyleClick() {
-    try {
-      mPickBorderStyle.accept(mBorderStyle);
-    } catch (Exception ignored) {
-      sLogger.w("Could not pick border style");
-    }
-  }
-
-  public void onBorderColorClick() {
-    try {
-      mPickColor.accept(Color.parseColor(mBorderColor));
-      mIsBorderColorPicking = true;
-    } catch (Exception ignored) {
-      sLogger.w("Could not pick border color");
-    }
-  }
-
-  public void onFillColorClick() {
-    try {
-      mPickColor.accept(Color.parseColor(mFillColor));
-      mIsBorderColorPicking = false;
-    } catch (Exception ignored) {
-      sLogger.w("Could not pick fill color");
-    }
-  }
-
-  public void onColorSelected(boolean positiveClick, int color) {
-    if (positiveClick) {
-      if (mIsBorderColorPicking) {
-        mBorderColor = colorToString(color);
-      } else {
-        mFillColor = colorToString(color);
-      }
-    }
-
-    notifyPropertyChanged(0);
-  }
-
-  public void onBorderStyleSelected(String borderStyle) {
-    mBorderStyle = borderStyle;
-    notifyPropertyChanged(0);
   }
 
   private void refreshCurrentEntity() {
@@ -208,22 +131,13 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
     }
   }
 
-  private String[] generateIconNames() {
-    String[] names = new String[mIcons.size()];
-    for (int i = 0; i < mIcons.size(); i++) {
-      names[i] = mIcons.get(i).getDisplayName();
-    }
-
-    return names;
-  }
-
   private void clearCurrentEntity() {
     mPoints.clear();
     mCurrentEntity = null;
   }
 
   private void setupDrawingMode(int newTabResource) {
-    mGGMapView.setOnMapGestureListener(mGestureLisetner);
+    mGGMapView.setOnMapGestureListener(mGestureListener);
 
     if (newTabResource == R.id.tab_point) {
       mOnMapClick = this::drawPoint;
@@ -239,7 +153,7 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
       mGGMapView.setOnMapGestureListener(null);
     }
 
-    notifyPropertyChanged(0);
+    notifyPropertyChanged(BR._all);
   }
 
   private void setupPointStyleVisibility() {
@@ -303,10 +217,6 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
     mMapDrawer.draw(mCurrentEntity);
   }
 
-  private String colorToString(int color) {
-    return "#" + Integer.toHexString(color).toUpperCase();
-  }
-
   private class OnDrawGestureListener implements OnMapGestureListener {
 
     @Override
@@ -319,6 +229,7 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
         mIsOnEditMode = true;
         mOnMapClick.accept(pointGeometry);
       } catch (Exception ignored) {
+        sLogger.w("Could not process map click");
       }
     }
   }
@@ -326,7 +237,9 @@ public class DynamicLayerEditViewModel extends BaseMapViewModel {
   private class OnPropertyChanged extends OnPropertyChangedCallback {
     @Override
     public void onPropertyChanged(Observable sender, int propertyId) {
-      refreshCurrentEntity();
+      if (propertyId == BR._all) {
+        refreshCurrentEntity();
+      }
     }
   }
 }
