+++ app/src/main/res/layout/fragment_send_geometry.xml
@@ -62,7 +62,7 @@
               android:id="@+id/border_style_picker"
               android:layout_width="@dimen/image_button_dimension_big"
               android:layout_height="@dimen/image_button_dimension_big"
-              android:onClick="@{(v) -> viewModel.onBorderStyleSelect()}"
+              android:onClick="@{(v) -> viewModel.onBorderStyleClick()}"
               android:padding="@dimen/spacing_small"
               app:srcCompat="@drawable/ic_border_style"
               />
@@ -71,7 +71,7 @@
               android:id="@+id/border_color_picker"
               android:layout_width="@dimen/image_button_dimension_big"
               android:layout_height="@dimen/image_button_dimension_big"
-              android:onClick="@{(v) -> viewModel.onBorderColorSelect()}"
+              android:onClick="@{(v) -> viewModel.onBorderColorClick()}"
               android:padding="@dimen/spacing_small"
               android:tint="@{viewModel.borderColor}"
               app:srcCompat="@drawable/ic_color_border"
@@ -81,7 +81,7 @@
               android:id="@+id/fill_color_picker"
               android:layout_width="@dimen/image_button_dimension_big"
               android:layout_height="@dimen/image_button_dimension_big"
-              android:onClick="@{(v) -> viewModel.onFillColorSelect()}"
+              android:onClick="@{(v) -> viewModel.onFillColorClick()}"
               android:padding="@dimen/spacing_small"
               android:tint="@{viewModel.fillColor}"
               app:srcCompat="@drawable/ic_color_fill"
