+++ circle.yml
@@ -17,7 +17,7 @@ machine:
 dependencies:
   pre:
     - chmod +x gradlew
-    - echo y | android update sdk --no-ui --all --filter "tools,platform-tools,android-23"
+    - echo y | android update sdk --no-ui --all --filter "tools,platform-tools,android-25"
     - echo y | android update sdk --no-ui --all --filter "build-tools-25.0.3"
     - echo y | android update sdk --no-ui --filter "extra-android-m2repository"
   cache_directories:
